+++ News_Title_Prediction/data_out/test_input.txt
@@ -98,10 +98,3 @@ GM 실적 최대
 이명박 문건 발견
 북한 핵시설 건설
 우병우 해외 도피
-
-
-
-
-
-
-
